perf: defer Studio analysis and below-fold editors
